chore: extract and test schedule calculations, cleanups
